requirements/dev.txt
Flask
flask_restful
flask_marshmallow
Flask-SQLAlchemy
pysqlite3
marshmallow-sqlalchemy